+++ src/app/courses/[courseId]/page.tsx
@@ -1,15 +1,10 @@
 import { notFound } from "next/navigation";
 import { eq } from "drizzle-orm";
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-} from "~/app/components/Card";
+import { Card, CardContent, CardHeader, CardTitle } from "~/components/Card";
 import { db } from "~/server/db"; // db should be your Drizzle db instance
 import { courses } from "~/server/db";
 import { SearchParams } from "next/dist/server/request/search-params";
-import { serverDetLang } from "~/app/utils/language";
+import { serverDetLang } from "~/utils/language";
 
 type CourseProps = {
   params: { courseId: string };
