+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "~/styles/globals.css";
 import { Suspense, type ReactNode } from "react";
 import { Providers } from "./providers"; // We'll create a separate file for theme and context providers
-import NavBar from "~/app/components/Navbar";
+import NavBar from "~/components/Navbar";
 import {
   ClerkProvider,
   SignInButton,
@@ -9,7 +9,7 @@ import {
   SignedOut,
   UserButton,
 } from "@clerk/nextjs";
-import { popScript } from "./tsScripts/generalScript";
+import { popScript } from "../tsScripts/generalScript";
 
 export const metadata = {
   title: "Open Uni Forum",
