+++ src/tsScripts/generalScript.ts
